script/baidu_translate/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Js2Py==0.70
pdfminer==20191125
pycryptodome==3.9.8
PyExecJS==1.5.1
pyjsparser==2.7.1
PyMySQL==0.10.1
pytz==2020.1
requests==2.24.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
